requirements.txt
asgiref==3.2.10
Django==3.1.2
django-ckeditor==6.0.0
django-cleanup==5.1.0
django-js-asset==1.2.2
gunicorn==20.0.4
Pillow==8.0.0
pytz==2020.1
sqlparse==0.4.1
whitenoise==5.2.0
